Brute_Force gui working properly Brute_Force fully functional
To do:
code needs cleaning/optimizing
add checkboxes to choose which lists i want included for the test

to enable multithreading:
consider breaking the code in two, one for the creation of the
entire Brute_Force list and one for the list web lookup.

allow both the creation and the lookup functions to have a continue
function to enable them to pick up where they left. needs storing info

create an optional third function to sort the brute force list